Move filter_compile/filter_run into standalone header
This will allow the functions to be used in other tests.
